Refactor: segmented user_catalog part 1
