+++ services/catalog/rs_server_catalog/app.py
@@ -27,7 +27,7 @@ from fastapi.responses import JSONResponse
 from fastapi.routing import APIRoute
 from httpx._config import DEFAULT_TIMEOUT_CONFIG
 from rs_server_catalog.data_lifecycle import DataLifecycle
-from rs_server_catalog.user_catalog import UserCatalog
+from rs_server_catalog.middleware.catalog_middleware import CatalogMiddleware
 from rs_server_common import settings as common_settings
 from rs_server_common.authentication.apikey import (
     APIKEY_AUTH_HEADER,
@@ -92,28 +92,11 @@ def add_parameter_owner_id(parameters: list[dict]) -> list[dict]:
     return parameters
 
 
-class UserCatalogMiddleware(BaseHTTPMiddleware):  # pylint: disable=too-few-public-methods
-    """The user catalog middleware."""
-
-    async def dispatch(self, request: Request, call_next: RequestResponseEndpoint) -> Response:
-        """Redirect the user catalog specific endpoint and adapt the response content."""
-        try:
-            response = await UserCatalog(api.client).dispatch(request, call_next)
-            return response
-        except (HTTPException, StarletteHTTPException) as exc:
-            phrase = HTTPStatus(exc.status_code).phrase
-            code = "".join(word.title() for word in phrase.split())
-            return JSONResponse(
-                status_code=exc.status_code,
-                content=ErrorResponse(code=code, description=str(exc.detail)),
-            )
-
-
 app: FastAPI = sfpg_app
 insert_middleware_after(
     app,
     BrotliMiddleware,
-    UserCatalogMiddleware,
+    CatalogMiddleware,
 )
 insert_middleware_after(app, CORSMiddleware, HandleExceptionsMiddleware)
 insert_middleware_after(
